requirements.txt
django 
python-dotenv
pytelegrambotapi
python-magic